requirements-dev.txt
pytorch-lightning
torchvision
hydra-core
optuna
